+++ vizro-core/src/vizro/models/_components/form/checklist.py
@@ -1,9 +1,10 @@
-from typing import Annotated, Literal, Optional
+from typing import Annotated, Any, Literal, Optional
 
 import dash_bootstrap_components as dbc
 from dash import html
 from pydantic import AfterValidator, Field, PrivateAttr, model_validator
 from pydantic.functional_serializers import PlainSerializer
+from pydantic.json_schema import SkipJsonSchema
 
 from vizro.models import VizroBaseModel
 from vizro.models._action._actions_chain import _action_validator_factory
@@ -16,8 +17,7 @@ class Checklist(VizroBaseModel):
     """Categorical multi-option selector `Checklist`.
 
     Can be provided to [`Filter`][vizro.models.Filter] or
-    [`Parameter`][vizro.models.Parameter]. Based on the underlying
-    [`dcc.Checklist`](https://dash.plotly.com/dash-core-components/checklist).
+    [`Parameter`][vizro.models.Parameter].
 
     Args:
         type (Literal["checklist"]): Defaults to `"checklist"`.
@@ -25,7 +25,11 @@ class Checklist(VizroBaseModel):
         value (Optional[MultiValueType]): See [`MultiValueType`][vizro.models.types.MultiValueType]. Defaults to `None`.
         title (str): Title to be displayed. Defaults to `""`.
         actions (list[ActionsType]): See [`ActionsType`][vizro.models.types.ActionsType]. Defaults to `[]`.
-
+        extra (Optional[dict[str, Any]]): Extra keyword arguments that are passed to `dbc.Checklist` and overwrite any
+            defaults chosen by the Vizro team. This may have unexpected behavior.
+            Visit the [dbc documentation](https://dash-bootstrap-components.opensource.faculty.ai/docs/components/input/)
+            to see all available arguments. [Not part of the official Vizro schema](../explanation/schema.md) and the
+            underlying component may change in the future. Defaults to `{}`.
     """
 
     type: Literal["checklist"] = "checklist"
@@ -40,6 +44,19 @@ class Checklist(VizroBaseModel):
         PlainSerializer(lambda x: x[0].actions),
         Field(default=[]),
     ]
+    extra: SkipJsonSchema[
+        Annotated[
+            dict[str, Any],
+            Field(
+                default={},
+                description="""Extra keyword arguments that are passed to `dbc.Checklist` and overwrite any
+            defaults chosen by the Vizro team. This may have unexpected behavior.
+            Visit the [dbc documentation](https://dash-bootstrap-components.opensource.faculty.ai/docs/components/input/)
+            to see all available arguments. [Not part of the official Vizro schema](../explanation/schema.md) and the
+            underlying component may change in the future. Defaults to `{}`.""",
+            ),
+        ]
+    ]
 
     _dynamic: bool = PrivateAttr(False)
 
@@ -52,16 +69,18 @@ class Checklist(VizroBaseModel):
     def __call__(self, options):
         full_options, default_value = get_options_and_default(options=options, multi=True)
 
+        defaults = {
+            "id": self.id,
+            "options": full_options,
+            "value": self.value if self.value is not None else [default_value],
+            "persistence": True,
+            "persistence_type": "session",
+        }
+
         return html.Fieldset(
             children=[
                 html.Legend(children=self.title, className="form-label") if self.title else None,
-                dbc.Checklist(
-                    id=self.id,
-                    options=full_options,
-                    value=self.value if self.value is not None else [default_value],
-                    persistence=True,
-                    persistence_type="session",
-                ),
+                dbc.Checklist(**(defaults | self.extra)),
             ]
         )