+++ vizro-core/src/vizro/models/_controls/parameter.py
@@ -8,8 +8,9 @@ from vizro.actions._parameter_action import _parameter
 from vizro.managers import model_manager
 from vizro.models import VizroBaseModel
 from vizro.models._components.form import Checklist, DatePicker, Dropdown, RadioItems, RangeSlider, Slider
+from vizro.models._controls._controls_utils import check_targets_present_on_page
 from vizro.models._models_utils import _log_call
-from vizro.models.types import SelectorType
+from vizro.models.types import ModelID, SelectorType
 
 
 def check_dot_notation(target):
@@ -17,13 +18,6 @@ def check_dot_notation(target):
         raise ValueError(
             f"Invalid target {target}. Targets must be supplied in the form <target_component>.<target_argument>"
         )
-    return target
-
-
-def check_target_present(target):
-    target_id = target.split(".")[0]
-    if target_id not in model_manager:
-        raise ValueError(f"Target {target_id} not found in model_manager.")
     return target
 
 
@@ -68,9 +62,8 @@ class Parameter(VizroBaseModel):
             Annotated[
                 str,
                 AfterValidator(check_dot_notation),
-                AfterValidator(check_target_present),
                 AfterValidator(check_data_frame_as_target_argument),
-                Field(description="Targets in the form `<target_component>.<target_argument>`."),
+                Field(description="Targets in the form of `<target_component>.<target_argument>`."),
             ]
         ],
         AfterValidator(check_duplicate_parameter_target),
@@ -79,6 +72,7 @@ class Parameter(VizroBaseModel):
 
     @_log_call
     def pre_build(self):
+        check_targets_present_on_page(control=self)
         self._check_numerical_and_temporal_selectors_values()
         self._check_categorical_selectors_options()
         self._set_selector_title()
@@ -104,7 +98,35 @@ class Parameter(VizroBaseModel):
             self.selector.title = ", ".join({target.rsplit(".")[-1] for target in self.targets})
 
     def _set_actions(self):
+        from vizro.models import Filter
+
         if not self.selector.actions:
-            self.selector.actions = [
-                _parameter(id=f"{PARAMETER_ACTION_PREFIX}_{self.id}", targets=self.targets)
+            page_dynamic_filters = [
+                filter
+                for filter in cast(
+                    Iterable[Filter], model_manager._get_models(Filter, page=model_manager._get_model_page(self))
+                )
+                if filter._dynamic
             ]
+
+            filter_targets: set[ModelID] = set()
+
+            # Extend parameter targets with dynamic filters linked to the same figure.
+            # Also, include dynamic filter targets to ensure that the new filter options are correctly calculated. This
+            # also ensures that the dynamic filter target is updated which is necessary if the filter value changes.
+            for figure in self.targets:
+                figure_id, figure_arg = figure.split(".", 1)
+                if figure_arg.startswith("data_frame"):
+                    for filter in page_dynamic_filters:
+                        if figure_id in filter.targets:
+                            filter_targets.add(filter.id)
+                            filter_targets |= set(filter.targets)
+
+            # Extending `self.targets` with `filter_targets` instead of redefining it to avoid triggering the
+            # pydantic validator like `check_dot_notation` on the `self.targets` again.
+            self.targets.extend(list(filter_targets))
+
+            # Ensure `Action.function["targets"]` remains consistent with `self.targets` by passing the updated
+            # `self.targets` directly to `_parameter`. This maintains synchronization between `self.targets` and
+            # `Action.function["targets"]`, preventing potential inconsistencies and confusion.
+            self.selector.actions = [_parameter(id=f"{PARAMETER_ACTION_PREFIX}_{self.id}", targets=self.targets)]