+++ vizro-core/src/vizro/models/_page.py
@@ -25,7 +25,7 @@ from vizro.models._action._actions_chain import ActionsChain, Trigger
 from vizro.models._layout import set_layout
 from vizro.models._models_utils import _log_call, check_captured_callable_model
 
-from .types import ComponentType, ControlType
+from .types import ComponentType, ControlType, FigureType
 
 # This is just used for type checking. Ideally it would inherit from some dash.development.base_component.Component
 # (e.g. html.Div) as well as TypedDict, but that's not possible, and Dash does not have typing support anyway. When
@@ -106,8 +106,18 @@ class Page(VizroBaseModel):
 
     @_log_call
     def pre_build(self):
+        # TODO-AV D 2: work out the best place to put this logic. It could feasibly go in _on_page_load instead.
+        #  Probably it's better where it is now since it avoid navigating up the model hierarchy
+        #  (action -> page -> figures) and instead just looks down (page -> figures).
+        #  Should there be validation inside _on_page_load to check that targets exist and are
+        #  on the page and target-able components (i.e. are dynamic and hence have _output_component_property)?
+        #  It's not needed urgently since we always calculate the targets ourselves so we know they are valid.
+        #  Similar comments apply to filter and parameter. Note that export_data has this logic built into the action
+        #  itself since the user specifies the target. In future we'll probably have a helper function like
+        #  get_all_targets_on_page() that's used in many actions. So maybe it makes sense to put it in the action for
+        #  on_page_load/filter/parameter too.
         figure_targets = [
-            model.id for model in cast(Iterable[VizroBaseModel], model_manager._get_models(FIGURE_MODELS, page=self))
+            model.id for model in cast(Iterable[FigureType], model_manager._get_models(FIGURE_MODELS, page=self))
         ]
         filter_targets = [
             filter.id
@@ -117,6 +127,7 @@ class Page(VizroBaseModel):
         targets = figure_targets + filter_targets
 
         if targets:
+            # TODO-AV2 A 3: can we simplify this to not use ActionsChain, just like we do for filters and parameters?
             self.actions = [
                 ActionsChain(
                     id=f"{ON_PAGE_LOAD_ACTION_PREFIX}_{self.id}",